Qaffee-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-RESTx==1.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
PyJWT==2.8.0
Werkzeug==2.3.7
SQLAlchemy==2.0.22
alembic==1.12.0
requests==2.31.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
oauthlib==3.2.2
stripe==8.4.0
bcrypt==4.1.2
gunicorn==21.2.0
python-dotenv==1.0.1
python-jose==3.3.0
email-validator==2.1.1
twilio==8.12.0
cloudinary==1.38.0 